TMA 807: Move ContextManager
