+++ lib/gooddata/bricks/middleware/context_manager.rb
