+++ spec/unit/bricks/middleware/context_manager_spec.rb
@@ -2,7 +2,7 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-require 'gooddata/core/core'
+require 'gooddata/bricks/middleware/context_manager'
 
 describe GoodData::ContextManager do
   it "Has GoodData::ContextManager class" do
